+++ changelog.md
@@ -3,7 +3,7 @@
 ## TBD
 
 - Added masteries for English, Iron Maid, Alligator, and Quarter Pounder
-- Added 6 new synergies
+- Added 7 new synergies
 - Updated required MtG API version to 1.7.7 to simplify non-active gun update logic in a few places
 - Added Electric damage type to Alligator projectiles
 - Increased Alligator projectile speed from 36 to 50
@@ -12,6 +12,7 @@
 - Added shader to visually indicate Alligator's charge level
 - Added impact sounds for Quarter Pounder projectiles
 - Fixed bad hand offset for one frame of Alligator reload animation
+- Removed some debug console output spat out when firing English
 
 ## 1.16.3 (2024-07-28)